Changes:
- Switched relevanceScore to double in DocumentSummary
- Updated InMemorySearchIndex to store separate term stats for title and content
- Implemented TF-based scoring with higher weight for matches in title
- Left TODO for IDF implementation

+++ src/main/java/com/github/stepanterkun/searchengine/search/domain/model/DocumentSummary.java
@@ -11,7 +11,7 @@ public record DocumentSummary(
         Long documentId,
         String documentTitle,
         DocumentStatus documentStatus,
-        int relevanceScore,
+        double relevanceScore,
         List<WordContextSnippet> wordSnippets
 ) {
 }
